create CppProgramRunner
