+++ src/prog_questions/utility/__init__.py
@@ -1,2 +1,3 @@
 from .CProgramRunner import CProgramRunner, ExecutionError, CompilationError, ExitCodeHandler, EnvironmentError, InternalError
 from .CommentMetric import CommentMetric
+from .CppProgramRunner import CppProgramRunner
